+++ src/mol-math/geometry/primitives/ray3d.ts
@@ -2,10 +2,12 @@
  * Copyright (c) 2025 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
+ * @author Gianluca Tomasello <giagitom@gmail.com>
  */
 
 import { Mat4 } from '../../linear-algebra/3d/mat4';
 import { Vec3 } from '../../linear-algebra/3d/vec3';
+import { Box3D } from './box3d';
 import { Sphere3D } from './sphere3d';
 
 interface Ray3D { origin: Vec3, direction: Vec3 }
@@ -40,6 +42,8 @@ namespace Ray3D {
         return out;
     }
 
+    //
+
     const tmpIR = Vec3();
     function _intersectSphere3D(ray: Ray3D, sphere: Sphere3D): number {
         const { center, radius } = sphere;
@@ -73,6 +77,59 @@ namespace Ray3D {
 
     export function isInsideSphere3D(ray: Ray3D, sphere: Sphere3D): boolean {
         return Vec3.distance(ray.origin, sphere.center) < sphere.radius;
+    }
+
+    //
+
+    function _intersectBox3D(ray: Ray3D, box: Box3D): number {
+        const { origin, direction } = ray;
+        const [minX, minY, minZ] = box.min;
+        const [maxX, maxY, maxZ] = box.max;
+        const [x, y, z] = origin;
+        const invDirX = 1.0 / direction[0];
+        const invDirY = 1.0 / direction[1];
+        const invDirZ = 1.0 / direction[2];
+        let tmin, tmax, tymin, tymax, tzmin, tzmax;
+        if (invDirX >= 0) {
+            tmin = (minX - x) * invDirX;
+            tmax = (maxX - x) * invDirX;
+        } else {
+            tmin = (maxX - x) * invDirX;
+            tmax = (minX - x) * invDirX;
+        }
+        if (invDirY >= 0) {
+            tymin = (minY - y) * invDirY;
+            tymax = (maxY - y) * invDirY;
+        } else {
+            tymin = (maxY - y) * invDirY;
+            tymax = (minY - y) * invDirY;
+        }
+        if ((tmin > tymax) || (tymin > tmax)) return -1;
+        if (tymin > tmin) tmin = tymin;
+        if (tymax < tmax) tmax = tymax;
+        if (invDirZ >= 0) {
+            tzmin = (minZ - z) * invDirZ;
+            tzmax = (maxZ - z) * invDirZ;
+        } else {
+            tzmin = (maxZ - z) * invDirZ;
+            tzmax = (minZ - z) * invDirZ;
+        }
+        if ((tmin > tzmax) || (tzmin > tmax)) return -1;
+        if (tzmin > tmin) tmin = tzmin;
+        if (tzmax < tmax) tmax = tzmax;
+        return tmin >= 0 ? tmin : -1;
+    }
+
+    export function intersectBox3D(out: Vec3, ray: Ray3D, box: Box3D): boolean {
+        const t = _intersectBox3D(ray, box);
+        if (t < 0) return false;
+
+        Vec3.scaleAndAdd(out, ray.origin, ray.direction, t);
+        return true;
+    }
+
+    export function isIntersectingBox3D(ray: Ray3D, box: Box3D): boolean {
+        return _intersectBox3D(ray, box) >= 0;
     }
 }